⬆️ improvement (App Frontend): Better Slack Notification with comparison branch: better-slack-notification
