+++ .gitignore
@@ -25,3 +25,4 @@ src/lib/paraglide
 
 # Tasks
 /tasksOut/*
+empheralData/*
